+++ qrap.c
@@ -111,7 +111,7 @@ void usage(const char *name)
  */
 int main(int argc, char **argv)
 {
-	struct timeval tv = { .tv_sec = 0, .tv_usec = 500 * 1000 };
+	struct timeval tv = { .tv_sec = 0, .tv_usec = (long)(500 * 1000) };
 	int i, s, qemu_argc = 0, addr_map = 0, has_dev = 0;
 	char *qemu_argv[ARG_MAX], dev_str[ARG_MAX];
 	struct sockaddr_un addr = {
